requirements.txt
streamlit==1.26.0
langchain==0.0.183
pypdf2
python-dotenv
openai
tiktoken
faiss-cpu
llama-cpp-python==0.1.55
streamlit-option-menu